suppresed last conversion warnings

+++ src/unittests/sys/test_mpi_handler.cpp
@@ -23,7 +23,7 @@ Test_start("mpi::",vector_allgather_ks,-1){
     {
         std::vector<u32> pointer_start_node(mpi_handler::world_size);
         pointer_start_node[0] = 0;
-        for(u32 i = 1; i < mpi_handler::world_size; i ++){
+        for(i32 i = 1; i < mpi_handler::world_size; i ++){
             pointer_start_node[i] = pointer_start_node[i-1] +5+ ((i-1)%5)*((i-1)%5);
         }
         pointer_start_node.push_back(mpi_handler::world_size*11);
@@ -38,7 +38,7 @@ Test_start("mpi::",vector_allgather_ks,-1){
     mpi_handler::vector_allgatherv_ks(local_vector, MPI_INT, recv_vec, MPI_INT, MPI_COMM_WORLD);
 
     //check correctness
-    for(u32 i = 1; i < mpi_handler::world_size*11; i ++) {
+    for(u32 i = 1; i < u32(mpi_handler::world_size)*11; i ++) {
         Test_assert(format("check_vec[%d] == recv_vec[%d]",i,i), check_vec[i] == recv_vec[i]);
     }
 
@@ -65,7 +65,7 @@ Test_start("mpi::",vector_allgather,-1){
     {
         std::vector<u32> pointer_start_node(mpi_handler::world_size);
         pointer_start_node[0] = 0;
-        for(u32 i = 1; i < mpi_handler::world_size; i ++){
+        for(i32 i = 1; i < mpi_handler::world_size; i ++){
             pointer_start_node[i] = pointer_start_node[i-1] +5+ ((i-1)%5)*((i-1)%5);
         }
         pointer_start_node.push_back(mpi_handler::world_size*11);
@@ -79,9 +79,9 @@ Test_start("mpi::",vector_allgather,-1){
     std::vector<u32> recv_vec;
     mpi_handler::vector_allgatherv(local_vector, MPI_INT, recv_vec, MPI_INT, MPI_COMM_WORLD);
 
-    Test_assert("vector lenght correct", recv_vec.size() == mpi_handler::world_size*11);
+    Test_assert("vector lenght correct", recv_vec.size() == u32(mpi_handler::world_size)*11);
     //check correctness
-    for(u32 i = 1; i < mpi_handler::world_size*11; i ++) {
+    for(u32 i = 1; i < u32(mpi_handler::world_size)*11; i ++) {
         Test_assert(format("check_vec[%d] == recv_vec[%d]",i,i), check_vec[i] == recv_vec[i]);
     }
 
@@ -114,7 +114,7 @@ Test_start("mpi::",sparse_alltoall,-1){
         std::uniform_int_distribution<u32> dist_send_cnt(0,10);
         std::uniform_int_distribution<u32> dist_node_id(0,mpi_handler::world_size-1);
 
-        for (u32 node_id = 0; node_id < mpi_handler::world_size; node_id++) {
+        for (i32 node_id = 0; node_id < mpi_handler::world_size; node_id++) {
             u32 send_cnt = dist_send_cnt(eng);
 
             arr_send_arr_node_id[node_id].resize(send_cnt);
@@ -176,7 +176,7 @@ Test_start("mpi::",sparse_alltoall,-1){
         u32 recv_node_tag = recv_arr_tag[recv_id];
         std::vector<float> recv_node_data = recv_arr_data[recv_id];
 
-        for(u32 sender_rank = 0; sender_rank < mpi_handler::world_size; sender_rank ++){
+        for(i32 sender_rank = 0; sender_rank < mpi_handler::world_size; sender_rank ++){
             for(u32 send_obj_id = 0; send_obj_id < arr_send_arr_node_id[sender_rank].size(); send_obj_id ++){
 
                 u32 send_node_id = sender_rank;
